Web Service/requirements.txt
Django==1.4.1
South==0.7.6
distribute==0.6.27
dj-database-url==0.2.1
django-piston==0.2.3
psycopg2==2.4.5
wsgiref==0.1.2
